Added ability to create multiple catalog entries at once.
This was actually added in this commit (https://github.com/dainst/arachnefrontend/commit/229ce84cb0203682cce902fdc02142e5a3673aa1) but I unfortunately messed up its message which is why it doesn't say so. This commit only adds some comments for the added lines to make it easier to revisit the code if necessary.