Add colored lines between places nodes according to chronological order
With this commit it works for single entities in the entity view. The necessary
date data is needed in the database and elastioc search index respectively.

The function filters out all the places which do not have a location information
 and also places with the same location twice right after another.

The colored lines shown are a gradient moving from green at the finding spot to
red at the last known location.